Calculations function complete
The calculations function finds the Altitude and Azimuth of a star at a certain location/time. The star's coordinates and time are hard coded in for now. My next step is to retrieve the coordinates of the star from a database, so any star can be found with a click of a button.